Fix: silence size_t->int narrowing warnings at BIO and header-token sites
- dkimf_loadkey(): cast keylen (size_t) to int for BIO_new_mem_buf();
  PEM/DER keys are always far below INT_MAX bytes.
- mlfi_eom(): cast strlen(p) to int for the local 'len' column counter;
  p is a strtok_r token from a DKIM_MAXHEADER-bounded buffer.

+++ opendkim/opendkim.c
@@ -4376,14 +4376,14 @@ dkimf_keytype(const char *keydata, size_t keylen)
 	if (strncmp(keydata, "-----", 5) == 0)
 	{
 		keylen = strlen(keydata);
-		bio = BIO_new_mem_buf(keydata, keylen);
+		bio = BIO_new_mem_buf(keydata, (int) keylen); /* key << INT_MAX */
 		if (bio == NULL)
 			return -1;
 		pkey = PEM_read_bio_PrivateKey(bio, NULL, NULL, NULL);
 	}
 	else
 	{
-		bio = BIO_new_mem_buf(keydata, keylen);
+		bio = BIO_new_mem_buf(keydata, (int) keylen); /* key << INT_MAX */
 		if (bio == NULL)
 			return -1;
 		pkey = d2i_PrivateKey_bio(bio, NULL);
@@ -12627,7 +12627,7 @@ mlfi_eom(SMFICTX *ctx)
 				     p = strtok_r(NULL, DELIMITER,
 				                  &last))
 				{
-					len = strlen(p);
+					len = (int) strlen(p); /* token bounded by DKIM_MAXHEADER */
 
 					if (!first)
 					{
